Clean up hasResponses endpoint on frontend
